+++ src/main/java/com/weshare/repository/UserRepository.java
@@ -3,7 +3,7 @@ import com.weshare.entity.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer>
+public interface UserRepository extends JpaRepository<User, Long>
 {
 	User findByUserNameAndPassword(String userName, String password);
 }
